Implement simple test program for mock_bluetooth.c
Create program that calls beagleblue_init, beagleblue_join and
beagleblue_exit to test there functionality. A make command was
created "testMockBluetooth" to create this program and a command
was added to "clean" to remove any executables that start with
test__ so that any future test programs can all be cleaned with the
same command. Update the gitignore to not include these test
programs.